backend/requirements.txt
# PDF Processing
PyMuPDF>=1.23.0

# API Communication
requests>=2.31.0

# Data Processing & Excel
pandas>=2.1.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Security
bcrypt>=4.1.0

# FastAPI Backend
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.18

# Image Enhancement
Pillow>=10.0.0
